RModel/delete

+++ app/controllers/PostController.php
@@ -103,6 +103,10 @@ class PostController extends BaseController
     public function actionView($topicId = null)
     {
         $topic = Topic::find($topicId)->join("group")->join("user")->first();
+        if($topic===null){
+            $this->page404();
+            return;
+        }
 
         $counter = $topic->increaseCounter();
         $commentTree = $topic->getComments();
@@ -219,8 +223,20 @@ class PostController extends BaseController
     public function actionDelete($topicId)
     {
         $topic = Topic::get($topicId);
+        if($topic===null)
+            return;
+
+        $counter = Counter::find("entityId",$topic->id)->find("entityType",Topic::ENTITY_TYPE)->first();
+        if($counter!=null)
+            $counter->delete();
+
+        // todo delete all rows at the same time
+        $comments = Comment::find("topicId",$topic->id)->all();
+        foreach($comments as $item)
+            $item->delete();
 
         $topic->delete();
+
         $this->flash("message", "Post " . $topic->title . " was deleted.");
         $this->redirect(Rays::referrerUri());
     }
